move divergent logic from blocks.js to wrappers
